Refactor extra API handlers into functions
